Add introductory kernel notebook and change style file path in notebooks
Main change is to add a notebook introducing the concept of a kernel for
those new to Gaussian processes. This focuses on mathematical intuition
and introduces useful concepts such as covariance matrices and
positive-definiteness.

Also noted that the README file for writing documentation hadn't been
updated since switching to using MkDocs, so updated it to reflect these
changes.

Also found that the relative path to the style file for the notebooks
caused issues when running the `poetry run mkdocs serve` command. Now
instead of using a relative path to the style file, we point to the URL
of the style file directly. I have copied the style file into the
`_static` directory, and a future PR will point to this URL instead.

Finally made a few other minor edits:
- Updated docstring for the periodic kernel as it was incomplete.
- Made some minor fixes to a few other docstrings I found flaws in.