Add classes for forming acquisition functions
Added an `AbstractAcquisitionFunctionBuilder` abstract base class for
forming acquisition functions, and added `ThompsonSampling` as a
concrete example. Unit tests have also been added.

Also renamed the `bayes_opt` module to `decision_making` to reflect the
fact that we aim for this module to support general sequential decision
making, and not solely Bayes opt.