Add classes for handling posteriors during the decision making loop
Added a `PosteriorHandler` class which can be used to obtain and update
posteriors during the decision making loop as the number of datapoints
in the training dataset changes (which is necessary for instantiating a
likelihood object).

Also added an `AbstractPosteriorOptimizer` abstract base class for
optimizing posterior hyperparameters during the decision making loop.
This allows for more customisability when optimising posteriors. Added a
concrete implementation in the form of an `AdamPosteriorOptimizer` which
can be used to optimize posteriors with the Adam optimizer provided by
optax.