Add logic for running a decision making loop
Added an `AbstractDecisionMaker` abstract base class which can be
implemented in order to define a decision making loop.

A concrete implementation, in the form of a `DecisionMaker` class,
has been added. At its heart it has two core methods:
1. `ask` which is used to get a point to be queried next.
2. `tell` which is used to tell the `DecisionMaker` about new
   observations. In a typical decision making setup this will result in
   the datsets and posteriors being updated.

In addition to this, in the `DecisionMaker` a `run` method is provided,
which will automatically run the decision making loop for n steps. After
the `ask` step, the functions in the `post_ask` list will be executed,
taking as arguments the decision maker and the point chosen to be
queried next. Similarly, after the `tell` step, the functions in
the `post_tell` list are executed, taking the decision maker as the sole
argument.